Reviewed-on: https://codeberg.org/jbruchon/libjodycode/pulls/1

+++ Makefile
@@ -42,12 +42,15 @@ ifeq ($(OS), Windows_NT)
 endif
 
 ifeq ($(UNAME_S), Darwin)
- LINK_OPTIONS += -Wl,-install_name,$(PROGRAM_NAME)$(SO_SUFFIX).$(API_VERSION)
+ ON_MACOS=1
+ LINK_OPTIONS += -Wl,-install_name,$(PROGRAM_NAME).$(API_VERSION)$(SO_SUFFIX)
  # Don't use unsupported compiler options on gcc 3/4 (Mac OS X 10.5.8 Xcode)
  GCCVERSION = $(shell expr `LC_ALL=C gcc -v 2>&1 | grep '[cn][cg] version' | sed 's/[^0-9]*//;s/[ .].*//'` \>= 5)
  STRIP_UNNEEDED = strip -S
  STRIP_DEBUG = strip -S
+ SO_SUFFIX = .dylib
 else
+ ON_LINUX=1
  LINK_OPTIONS += -Wl,-soname,$(PROGRAM_NAME)$(SO_SUFFIX).$(API_VERSION)
  GCCVERSION = 1
  STRIP_UNNEEDED = strip --strip-unneeded
@@ -133,9 +136,12 @@ all: sharedlib staticlib
 	-@test "$(CROSS_DETECT)" = "cross" && echo "NOTICE: SIMD disabled: !x86_64 or a cross-compiler detected (CC = $(CC))" || true
 
 sharedlib: $(OBJS) $(SIMD_OBJS)
-	$(CC) -shared -o $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(OBJS) $(SIMD_OBJS) $(LDFLAGS) $(CFLAGS) $(CFLAGS_EXTRA)
-	-test "$(ON_WINDOWS)" != "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
-	-test "$(ON_WINDOWS)" != "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(PROGRAM_NAME)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" != "1" && $(CC) -shared -o $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(OBJS) $(SIMD_OBJS) $(LDFLAGS) $(CFLAGS) $(CFLAGS_EXTRA)
+	-test "$(ON_MACOS)" == "1" && $(CC) -shared -o $(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX) $(OBJS) $(SIMD_OBJS) $(LDFLAGS) $(CFLAGS) $(CFLAGS_EXTRA)
+	-test "$(ON_LINUX)" == "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
+	-test "$(ON_LINUX)" == "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(PROGRAM_NAME)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" == "1" && $(LN) $(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX) $(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" == "1" && $(LN) $(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX) $(PROGRAM_NAME)$(SO_SUFFIX)
 
 staticlib: $(OBJS) $(SIMD_OBJS)
 	$(AR) rcs libjodycode$(LIB_SUFFIX) $(OBJS) $(SIMD_OBJS)
@@ -182,10 +188,13 @@ installdirs:
 	test -e $(DESTDIR)$(MAN7_DIR) || $(MKDIR) $(DESTDIR)$(MAN7_DIR)
 
 installfiles:
-	$(INSTALL_PROGRAM) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)            $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME).$(SO_SUFFIX)$(VERSION)
-	-test "$(ON_WINDOWS)" != "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
-	-test "$(ON_WINDOWS)" != "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)
-	$(INSTALL_DATA) $(PROGRAM_NAME)$(LIB_SUFFIX  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(LIB_SUFFIX)
+	-test "$(ON_MACOS)" != "1" && $(INSTALL_PROGRAM) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)
+	-test "$(ON_MACOS)" == "1" && $(INSTALL_PROGRAM) $(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX)
+	-test "$(ON_LINUX)" == "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
+	-test "$(ON_LINUX)" == "1" && $(LN) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" == "1" && $(LN) $(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" == "1" && $(LN) $(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX) $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)
+	$(INSTALL_DATA) $(PROGRAM_NAME)$(LIB_SUFFIX)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(LIB_SUFFIX)
 	$(INSTALL_DATA) $(PROGRAM_NAME).h  $(DESTDIR)$(INC_DIR)/$(PROGRAM_NAME).h
 	$(INSTALL_DATA) $(PROGRAM_NAME).7  $(DESTDIR)$(MAN7_DIR)/$(PROGRAM_NAME).7
 
@@ -197,8 +206,10 @@ uninstalldirs:
 	-test -e $(DESTDIR)$(MAN7_DIR) && $(RMDIR) $(DESTDIR)$(MAN7_DIR)
 
 uninstallfiles:
-	$(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)
-	$(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
+	-test "$(ON_MACOS)" != "1" && $(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)
+	-test "$(ON_MACOS)" != "1" && $(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR)
+	-test "$(ON_MACOS)" == "1" && $(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX)
+	-test "$(ON_MACOS)" == "1" && $(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX)
 	$(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(SO_SUFFIX)
 	$(RM)  $(DESTDIR)$(LIB_DIR)/$(PROGRAM_NAME)$(LIB_SUFFIX)
 	$(RM)  $(DESTDIR)$(INC_DIR)/$(PROGRAM_NAME).h
@@ -217,7 +228,8 @@ objsclean:
 	$(RM) $(OBJS) $(SIMD_OBJS) vercheck.o *.obj
 
 clean: objsclean
-	$(RM) $(PROGRAM_NAME)$(SO_SUFFIX) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)
+	-test "$(ON_MACOS)" != "1" && $(RM) $(PROGRAM_NAME)$(SO_SUFFIX) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION_MAJOR) $(PROGRAM_NAME)$(SO_SUFFIX)$(VERSION)
+	-test "$(ON_MACOS)" == "1" && $(RM) $(PROGRAM_NAME)$(SO_SUFFIX) $(PROGRAM_NAME)$(VERSION_MAJOR)$(SO_SUFFIX) $(PROGRAM_NAME)$(VERSION)$(SO_SUFFIX)
 	$(RM) $(PROGRAM_NAME)$(LIB_SUFFIX) apiver cacheinfo vercheck
 	$(RM) *~ helper_code/*~ libjodycode.so.* libjodycode.dll.* .*.un~ *.gcno *.gcda *.gcov
